corrections to the list of comments
