+++ src/main/java/ru/mail/knhel7/money_transfer_service/model/Money.java
@@ -1,12 +1,19 @@
 package ru.mail.knhel7.money_transfer_service.model;
 
 import jakarta.validation.constraints.Min;
-import ru.mail.knhel7.money_transfer_service.model.addit_code.Currency;
 
-public record Money(Currency currency, @Min(value = 0, message = "should not be negative") Integer value) {
+public record Money(String currency , @Min(value = 0, message = "should not be negative") Integer value) {
 
-    public Money(Currency currency, Integer value) {
-        this.currency = currency;
+    public Money() {
+        this(0);
+    }
+
+    public Money(Integer value) {
+        this("RUR", value);
+    }
+
+    public Money(String currency, @Min(value = 0, message = "should not be negative") Integer value) {
+        this.currency = "RUR";
         this.value = value;
     }
