+++ src/main/java/ru/mail/knhel7/money_transfer_service/model/transaction/Transaction.java
@@ -2,21 +2,20 @@ package ru.mail.knhel7.money_transfer_service.model.transaction;
 
 import lombok.Getter;
 import lombok.Setter;
-import ru.mail.knhel7.money_transfer_service.model.addit_code.TransactionStatus;
-import ru.mail.knhel7.money_transfer_service.tools.DateTimeTool;
+import ru.mail.knhel7.money_transfer_service.util.DateTimeUtil;
+
+import java.util.concurrent.atomic.AtomicInteger;
 
 
 @Getter
-public class Transaction<T> implements ITransaction<T>{
+public class Transaction<T> {
 
-    private static Integer Counter = 1;
+    private static final AtomicInteger Counter = new AtomicInteger(1);
 
-    private final Integer ID = Counter++;
-    private final String date = DateTimeTool.getDate();
-    private final String time = DateTimeTool.getTime();
+    private final Integer ID = Counter.getAndIncrement();
+    private final String date = DateTimeUtil.getDate();
+    private final String time = DateTimeUtil.getTime();
 
-    @Setter
-    private TransactionStatus status = TransactionStatus.NEW;
     @Setter
     private  String comment = "Зарегистрирована новая операция" + date + " " + time;
