+++ src/test/java/ru/mail/knhel7/money_transfer_service/TransferRepositoryTests.java
@@ -1,17 +1,10 @@
 package ru.mail.knhel7.money_transfer_service;
 
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-import ru.mail.knhel7.money_transfer_service.repository.CardRepo;
-
-import java.util.Optional;
+import ru.mail.knhel7.money_transfer_service.repository.CardRepoImpl;
 
 public class TransferRepositoryTests {
 
-    private CardRepo repo;
+    private CardRepoImpl repo;
