+++ src/test/java/ru/mail/knhel7/money_transfer_service/TransferServiceTests.java
@@ -1,10 +1,10 @@
 package ru.mail.knhel7.money_transfer_service;
 
-import ru.mail.knhel7.money_transfer_service.service.CardMoneyTransferService;
+import ru.mail.knhel7.money_transfer_service.service.CardMoneyTransferServiceImpl;
 
 public class TransferServiceTests {
 
-    private CardMoneyTransferService service;
+    private CardMoneyTransferServiceImpl service;
